requirements.txt
google-adk
google-genai
aiosqlite
greenlet
jupyter
ipykernel

